misc: sec_jack: detect presence of headset at probe time
If a headset is connected before the kernel is booted, no interrupt
will occur and the headset will not be detected. The contents of
the threaded IRQ handler are moved into a work queue so it can be
queued at probe() time. The threaded IRQ handler is no longer
required and the existing single threaded work queue is shared.

Change-Id: Ia528842501c15de41102dd652a60e520add7e456
Signed-off-by: Simon Wilson <simonwilson@google.com>