AppArmor: Fix oops in policy unpack auditing
Post unpacking of policy a verification pass is made on x transition
indexes.  When this fails a call to audit_iface is made resulting in an
oops, because audit_iface is expecting a valid buffer position but
since the failure comes from post unpack verification there is none.

Make the position argument optional so that audit_iface can be called
from post unpack verification.

Change-Id: I6a984ac83f83f8909d5e1f5317a3e2916b6fd33f
Signed-off-by: John Johansen <john.johansen@canonical.com>