sync: Fix race condition between merge and signal
The copied sync_pt was activated immediately. If the sync_pt was
signaled before the entire merge was completed, the new fence's pt_list
could be iterated over while it is still in the process of being
created.

Moving the the sync_pt_activate call for all new sync_pts to after both
the sync_fence_copy_pts and the sync_fence_merge_pts calls ensure that
the pt_list is complete and immutable before it can be reached from the
timeline's active list.

Change-Id: Ib13c7323c7cca5e4916498b12636e184606a8dcb
Signed-off-by: Ørjan Eide <orjan.eide@arm.com>
Signed-off-by: Erik Gilling <konkers@android.com>